requirements.txt
blinker~=1.7.0
click~=8.1.7
flask~=3.0.0
gunicorn~=21.2.0
importlib-metadata~=6.8.0
itsdangerous~=2.1.2
Jinja2~=3.1.2
MarkupSafe~=2.1.3
mysql-connector-python~=8.2.0
packaging~=23.2
protobuf~=4.21.12
python-dotenv~=1.0.0
werkzeug~=3.0.1
zipp~=3.17.0
